WIP Migrate replacer to edit
